pyproject.toml
[project]
name = "kajima-mailroom"
version = "0.1.0"
description = "Air-gapped AI document classification appliance for local councils"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=4.0.0",
    "flask>=3.0.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0",
    "httpx>=0.27.0",
    "rich>=13.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
mailroom = "main:main"
